% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/favnums.R
\docType{package}
\name{favnums}
\alias{favnums}
\alias{favnums-package}
\title{A Dataset of Favourite Numbers}
\description{
This package provides a dataset of favourite numbers, selected from an online poll of over
30,000 people by Alex Bellos
}
\seealso{
\code{\link{favourite_numbers}} and the \href{http://pages.bloomsbury.com/favouritenumber}{original dataset}.
}

