% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/favnums.R
\docType{data}
\name{favourite_numbers}
\alias{favourite_numbers}
\title{Favourite Numbers based on an online poll}
\format{A data frame with 1123 rows and 4 variables:
\describe{
  \item{number}{the actual number chosen. May be NA in the case of "imaginary" numbers, or Infinite.}
  \item{frequency}{the number of times this number was chosen.}
  \item{percentage}{the percentage of user answers a particular number represents.}
  \item{description}{descriptions of the number's importance, as provided by Alex Bellos. Often NA.}
}}
\source{
\url{http://pages.bloomsbury.com/favouritenumber}
}
\usage{
favourite_numbers
}
\description{
A dataset containing the favourite numbers selected by over
30,000 people in an online poll.
}
\keyword{datasets}

